requirements.txt
annotated-types==0.7.0
cx-oracle==8.3.0
dnspython==2.6.1
greenlet==3.0.3
motor==3.5.0
numpy==2.0.0
pandas==2.2.2
pip==23.0.1
pydantic-core==2.20.1
pydantic==2.8.2
pymongo==4.8.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==65.5.0
six==1.16.0
sqlalchemy==2.0.31
typing-extensions==4.12.2
tzdata==2024.1